:recycle: Put import statements only when needed instead globally
It appears that the version of Jupyter or the Python extension for
VS Code did not want to import the respective libraries at the top
of the file. Instead, the imports are now located in the cells when
they are first needed. VS Code seems to be fine with this approach.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>