:alembic: Update dataset for training
The previous dataset that was used was Rachael Tatman's English word
frequency dataset. However, there were some issues pertaining to this
dataset that could have potentially thrown the network off. Some of
the words in the dataset were proper names that weren't even English,
such as 'Koprivshtitsa'. This made it hard to tell what was going on,
and it made the dataset unnecessarily large.

As a result, the new dataset will use the Japanese words and the words
file found on UNIX systems. For the case of this repository, a copy of
the words file (/usr/share/dict/words) from a Pop!_OS 20.04 LTS (based
on Ubuntu Linux) is provided. This has sped up the network by a large
margin, but this doesn't seem to affect accuracy.

Some other experiments have been done to change the accuracy:

- Changing one of the Dense layers to twice the amount of the original
  inputs.
- Changing the means of normalization.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>