:alembic: Adjust networking parameters
The network paramaters have been adjusted to try and optimize the
accuracy, with some success.

First, the dataset has been reshaped to rely on the average word
length, rather than the maximum. This allows for entries that are
not entirely filled with zeroes.

Additionally, the training and testing sets will now filter out
words that are larger than the average length.

The batch size of the network has been reduced to 64, given that the
dataset is much smaller and doesn't need as big of a batch to train
on.

Finally, the method of detecting loss has been changed to binary
cross entropy instead of categorical cross entropy.

Although the accuracy has not changed significantly, I have observed
the following:

- The loss is now significantly dropping compared to before, where
  it was completely stagnant.
- The network now does filter out words that don't match its
  predictions.

The next commits will try to focus on the GAN itself, and researching
why this is difficult when it comes to text.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>