:recycle: Use preprocessing utilities from RNN labs
In prevous commits, a function fill_word_features was used to pad
words to a specified length. This is unnecessary since Keras has a
utility designed specifically for this: pad_sequences. The notebook
has been refactored to use this tool instead of the
fill_word_features utility.

Additionally, the random shuffling is now handled by the test/train
split, instead of shuffling beforehand.

Finally, the early stopping callback has been removed from the fit
call to show the overfitting issue present with the current
implementation.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>